(split)bug: 92435
Group related issues when using CustomLdapProvisioning.
When computing ZimbraBaseDN for CustomLdapDIT, it could potentially have a trailing ","
CustomLdapDIT filterGroupsByDomain and filterDistributionListsByDomain now return valid filters
instead of throwing UnsupportedOperationException.  Assumption (which seems valid from previous comment)
is that either "mail" or "zimbraMailAlias" will contain the mail address for the DL/group
CustomLdapProvisioning overrides getAllDistributionLists(Domain) and getAllGroups(domain)
The code in LdapProvisioning assumes that these are stored in the subtree under the domain - which isn't a
valid assumption for CustomDIT.
getCustomDynamicGroupMembership is also over-ridden on the assumption that dynamic groups are not supported.

Tested these commands which work in a CustomLdapProvisioning environment where there were problems before:
    zmsoap -t account -m user2@domain -p test123 GetInfoRequest @rights=createDistList
    zmprov gadl
    zmprov getGrants -g usr user2@domain 1  # will now show grants inherited via DL
    zmprov gam user2@domain

https://reviewboard.eng.zimbra.com/r/1666/

http://bugzilla.zimbra.com/show_bug.cgi?id=92435

Copied from Perforce
 Change: 514921