bug:92900
enhanced JAXB for SOAP Header Context used in requests.
Based on code in ZimbraSoapContext constructor and SoapTransport.generateSoapMessage

One limitation.  Doesn't support attributes on <authToken>.  This was poorly designed - using the text body of <authToken> as the value but also allowing sub-elements for attributes leads to major usability issues in JAXB.   (You can't have @XmlElement and @XmlValue in the same class...)  If it becomes important in the future, suggest we allow specifying the authToken value in an attribute as an alternative and requiring that when using attributes on the auth token.
https://reviewboard.eng.zimbra.com/r/1871/

http://bugzilla.zimbra.com/show_bug.cgi?id=92900

Copied from Perforce
 Change: 516775