(split)bug:93114,93399
Potential regex related DOS attack vectors using an authorized account.
93114 - pathalogical IMAP LIST commands.
        New zimbraRegexMaxAccessesWhenMatching Config attribute added
        Also do a pre-optimization to collapse repeating * and % chars which should mitigate
        the issue seen by the originally reporting customer where LIST was being performed
        blindly with the name of a mailbox which contained repeating * chars
93399 - BrowseRequest regex expression used blindly
        Uses AccessBoundedRegex with hard coded MAX_REGEX_ACCESSES = 100000.  Not made
        configurable as believe that any real world regex expressions in this context would
        not be very resource intensive and also failing results are likely to be addressable
        by the customer (less likely to be true with IMAP clients)

New AccessBoundedRegex class (and helper class AccessBoundedCharSequence) allow a mechanism
for limiting how much resource is expended computing the regex.  The limit maps to the
number of calls to charAt(int index) in the target CharSequence when performing a match.
An alternative approach of using a Future with a timeout was considered but this mechanism
was prefered because it ensures consistent behavior and doesn't require additional threads.

TestImap updated to be compatible with RunUnitTests
https://reviewboard.eng.zimbra.com/r/1904/

http://bugzilla.zimbra.com/show_bug.cgi?id=93114
http://bugzilla.zimbra.com/show_bug.cgi?id=93399

Copied from Perforce
 Change: 517288