(split)bug:93268
zimbraFreebusyExchangeAuthPassword should return VALUE-BLOCKED for SOAP calls by default
Affected responses:
GetAllConfigResponse, GetEffectiveRightsResponse, GetDomainResponse, GetCosReponse, GetAccountResponse

In the Admin Console for Global Config and for Domains, there is a "Check the settings" function.
To ensure reasonable usability, VALUE-BLOCKED from SOAP responses is changed to "" so that it is
clear that the password needs to be typed in again.
https://reviewboard.eng.zimbra.com/r/1950/

http://bugzilla.zimbra.com/show_bug.cgi?id=93268

Copied from Perforce
 Change: 517820