(split)bug 92718 - only deploy zimlets on mail client servers
This consolidates the logic for determining the deployable servers.
"Deployable" in this instance means that the server either:

a) the current server running the "zimlet" servlet
b) runs both the "zimlet" and "service" servlet

This means that you have to log into each UI server directly using SSH
and deploy zimlets manually.

http://bugzilla.zimbra.com/show_bug.cgi?id=92718

Copied from Perforce
 Change: 518204