(split)Integrated the following from //depot/zimbra/JP-INTERNS/...: Change 514279 by gprasad@gprasad-mac-jpinterns on 2014/06/13 12:32:13
bug: 77448


Problem : GalSyncAccount contacts were not being removed when datasource was removed using zmprov deleteDataSource. 
Analysis: The mailbox associated with the gal account wasn't being removed on deleteDatasource
Fix : Added code to remove the mailbox associated with the gal account on deleteDatasource.
Also observed that zmgsautil deleteAccount was not removing the mailbox associted with the account and it remained obselete , so made the similar changes to DeleteGalSyncAccount.java to fix this issue aswell.

http://bugzilla.zimbra.com/show_bug.cgi?id=77448

Copied from Perforce
 Change: 520035