Bug: 86304 - bad attachment causes HTML conversion error and heap inflation on preview
Following are the converted logs for the preview request for file PLANEJAMENTO EFO-2407.xlsx attached with the bug.

[Fri Aug 22 01:26:28.839704 2014] [:info] [pid 31454] mod_convert: clean dir=/opt/zimbra/convertd/convert/convert
[Fri Aug 22 01:26:28.856366 2014] [:info] [pid 31454] mod_convert: clean dir=/opt/zimbra/data/tmp/convert
[Fri Aug 22 01:26:47.252761 2014] [:info] [pid 31454] mod_convert: post="PLANEJAMENTO EFO-2407.xlsx" i=41856 o=540699723 fmt=h ms=18225

In the logs, you can see that the input file size is approximately 41 KB
but the response received from convertd is very large approximately 520 MB.
Coverting this response into a byte array or a String results in JAVA Heap Space OOME.
I am not sure whether the large response size is because of some bug on the convertd side. But, the following work around will help in avoiding the JAVA Heap space OOME.

The work around is to process the response only if it is smaller than a predefined limit.

1. Define DebugConfig.convertdMaxResponseSize as the default max response size (20 MB)
2. In the constructor set maxResponseSize to twice the value of the attribute ?zimbraFileUploadMaxSize?
3. If ?zimbraFileUploadMaxSize? is zero (which means no limit), then set maxResponseSize = DebugConfig.convertdMaxResponseSize
4. In call(HttpMethod m), process the response only if the response size is smaller than maxResponseSize else throw an exception with message ?Response is too large?. 
5. OOME will not occur and on the preview UI, the user will see the option to download the file instead of file preview.

http://bugzilla.zimbra.com/show_bug.cgi?id=86304

Copied from Perforce
 Change: 521733