Integrate change 521937 from JUDASPRIEST bug: 24436
Add new attribute zimbraPrefDelegatedSendSaveTarget to control where sent messages are saved when sent from via a sendAs or sendOnBehalfOf grant. The possible values are 'owner', 'sender', 'both', or 'none'; with 'owner' as the default. 

This setting is read from the owner (aka granter) and is only applicable if the owner account also has zimbraPrefSaveToSent enabled.

http://bugzilla.zimbra.com/show_bug.cgi?id=24436

Copied from Perforce
 Change: 521938