bug 95219 - Distributed caching abstraction
Replaced DB-based reloading of MailboxData (stats) that was introduced for Always On phase 1 with a cache interface and Local, Memcached, and Redis adapters. No cache adapter is currently configured, and the product will avoid cache checks in this case, which prevents any new code from entering the execution path at this point. The local adapter is only used for unit tests. The distributed cache adapters will be configurable in a future release.

Fixes ZimbraMemcachedClient::isConnected(), which was always returning true. Now it returns true only if at least an initial connection was established.

http://bugzilla.zimbra.com/show_bug.cgi?id=95219

Copied from Perforce
 Change: 521945