(split)bug 95482
Purpose:

In order to attach files as ZSS links to ZCS emails, ZCS has to make call to ZSS to assign permissions on files to recipients. The call has to be made at about the same time as a message with links is being sent. To make the call, the extension needs access to list of recipients, ZSS URLs of files and map of permissions to be assigned to each ZSS URL

New code:
Add SendMailListener interface and a ConcurrentHashMap of SendMailListeners to register/deregister listeners for MailSender
Call SendMailListener::handle method of each listener right after sending a message via SMTP

New tests:
 - test that a listener is being triggered
 - test that a listener is being triggered only once per message
 - test that a listener is able to retrieve custom mime headers 

http://bugzilla.zimbra.com/show_bug.cgi?id=95482

Copied from Perforce
 Change: 523829