(split)bug 86944 - Always On Architecture (Distributed Datastore)
Remove ZooKeeper & Curator.

This change backs out CL477281, which introduced some ZooKeeper-held state relating to servers being manually marked active/online or inactive/offline. This state has no use or meaning yet in the present system (it isn't used anywhere AFAIKT), and does also not blend with the nginx-style of dynamic (non-manual) upstream status determination that we have in mind now for Always On.

This change also backs out parts of CL472800 and CL474800 relating to NginxLookup use of ZooKeeper for service discovery.

This change retires the ZooKeeper-based distributed MailboxLock, which is not replaced with anything yet at this point, but is being addressed for 9.0 by bug 85257 (Distributed Mailbox Lock management).

http://bugzilla.zimbra.com/show_bug.cgi?id=86944

Copied from Perforce
 Change: 525176