(split)Bug 95219 - Distributed caching abstraction.
Replace memcached-based ItemCache from Always On phase 1 with a pure interface plus Local, Memcached, and Redis adapters.

Unlike prior commits, this one does not include integration of the mailstore to make use of the new abstraction. That is held up by the current MailboxChange design, which currently passes a full snapshot of the item cache into the CurrentChange object, so that for the duration of a transaction a stable cache snapshot is present. This is obviously easy to do with a Java collection, but less easy to do with Redis, and very hard or impossible with Memcached. So the integration is deferred to a later release, until a purely stateless mailstore would require it.

In the meanwhile, the adapters will be available in the codebase, and have solid test coverage. And in the meanwhile, the Local adapter will continue to be used, and a pub/sub adapter is used by the Session and MailboxListener to perform ItemCache invalidations (this has been true since 8.5 or Always On phase 1).

http://bugzilla.zimbra.com/show_bug.cgi?id=95219

Copied from Perforce
 Change: 525560