(split)Bug: 84469 - zmmailbox: duplicate mailbox items after incremental import
1. Overloaded method addMessage() in Mailbox.java so that the existing blob digest 
can be used in case of imports.

The bugs seems to be occuring in case of imports from older version. The digest created during import is different from the digest in the meta file of the mail item. So, the current code is not able to identify duplicates if the user tries to import the same tar/tgz file again. I think this might be because of the change in java version. 
Reproduced the bug with the tar attached in the bug. I was also able to reproduce this bug
with data imported from Zimbra 7.2.7 running on zqa-418.eng.zimbra.com

Testing done - 
1. Import using the attachment in the bug - 
    - Able to repro the bug
    - There are no duplicates after applying the fix

2. Import from Zimbra 7.2.7 running on zqa-418.eng.zimbra.com
    - Able to repro the bug
    - There are no duplicates after applying the fix

3. Import from zqa-061.eng.zimbra.com
    - Not able to repro the bug

http://bugzilla.zimbra.com/show_bug.cgi?id=84469

Copied from Perforce
 Change: 526647