bug:96145
zmtzdata2ical generates incorrect VTIMEZONE for Europe/Moscow and others where rule has only just come into effect.
Problem was that the Day type used is multipurpose and the compareTo method doesn't work correctly for ON comparisons, so, if the reference date for the end of relevance for a rule is recent (within the same month) we incorrectly think it still applies.
https://reviewboard.eng.zimbra.com/r/2869/

http://bugzilla.zimbra.com/show_bug.cgi?id=96145

Copied from Perforce
 Change: 527206