(split)bug:95440
ActiveSync originated timezones not recognized by Yosemite Calendar can sometimes result in Yosemite PUTing an update to the calendar using the odd TZID but NOT including the corresponding VTIMEZONE.

To avoid this try harder to ensure that timezones we expose via CalDAV are cannonical ones.  Also cull unused VTIMEZONEs.
As the rules in the ActiveSync timezones can be a bit buggy as to what hour timezone transitions take place, relax the rules more than in the past when looking for a match.

Improved tracing available for GET

https://reviewboard.eng.zimbra.com/r/2961/

http://bugzilla.zimbra.com/show_bug.cgi?id=95440

Copied from Perforce
 Change: 528116