(split)Bug: 84727 - Importing a message under a folder whose name is in Japanese causes IllegalArgumentException
The exception was occuring because correct charset was not being used for reading folder_ja.zip

As per the solution for JDK bug JDK-4244499 : ZipEntry() does not convert filenames from Unicode to platform,
We are now using the ZipInputStream constructor - ZipInputStream(InputStream, Charset).
This will ensure that a correct Charset is used for reading a ZipInputStream

To fix the bug we also need to make changes in Ajax Client.
For the import functionality, the value of request param "charset" is always UTF-8.
In case of japanese it can be either windows-31j or Shift_JIS. Ajax client should send the correct value in request param.
Should we open a new bug for this?

Testing done - 

folder_ja.zip is imported successfully using the following curl commands -

1. curl -k -u admin:test123 --data-binary @/Users/rambasta/dev/bug-84727/folder_ja.zip "https://localhost:7071/service/home/   user1@host.local/?fmt=zip&resolve=skip&charset=windows-31j"

2. curl -k -u admin:test123 --data-binary @/Users/rambasta/dev/bug-84727/folder_ja.zip "https://localhost:7071/service/home/user3@host.local/?fmt=zip&resolve=skip&charset=Shift_JIS"

For more information on the jdk bug please refer the following link - 
https://blogs.oracle.com/xuemingshen/entry/non_utf_8_encoding_in

https://bugzilla.zimbra.com/show_bug.cgi?id=84727

Copied from Perforce
 Change: 530464