(split)Bug: 86938
This bug is caused by an unhandled IOException thrown by a corrupt lucene index. When an index is corrupt, the LuceneIndex class will throw an IOException when opening a searcher or an indexer. While methods that call openIndexer() handle this exception gracefully, ensuring that message delivery does not get interrupted by a failed indexing step, some calls to openSearcher() propagate the exception. In most cases (for example, during a maibox search) this is the right thing to do. However, in the case of sending a message, an exception thrown by openSearcher() should not interfere with message delivery.

The solution for this bug is simply to handle Lucene IOExceptions in the message delivery code path. Specifically this exception can be thrown by MailboxIndex.existsInContacts, which tries to open a searcher. When the enlosing method, Mailbox.newContactAddrs, encounters exception, it will now log an error but will not propagate it.

Tested by resetting the mailbox, manually corrupting the lucene index files, and sending a message. An exception is logged, but the message is delivered.

Mailbox.java:
 - updated newContactAddrs method to catch an IOException thrown by MailboxIndex.existsInContacts, logging the error.

https://bugzilla.zimbra.com/show_bug.cgi?id=86938

Copied from Perforce
 Change: 530849