(split)Bug: 95991
Updated ZimbraServer pom.xml to bind setup of the embedded SOLR server directory into the generate-test-resources lifecycle. Not sure if I made all the necessary adjustments, but "mvn clean install" succeeds from both main/ and main/ZimbraServer.

Also, updated MailboxTestUtil to not fail when unable to find files when calling setupEmbeddedSolrDirs(). We want to be able to run unit tests outside of maven, so this way of creating the SOLR directory is still necessary; we just don't want it to break when maven is running things.

https://bugzilla.zimbra.com/show_bug.cgi?id=95991

Copied from Perforce
 Change: 531038