bug: 86249
Use fuzzy matching for timezones more to reduce the number of buggy timezone definitions entering the system.
There are frequent issues with timezone definitions having the wrong onset/offset times, so, if we don't find an exact match, it makes sense to accept a zone which only differs in the onset times (bit with the correct onset days)

With the test message from the bug.  Before this change the CalDAV GET of the calendar item created during delivery had "Europe/Berlin" as the TZID rather than "GMT +0100 (Standard) / GMT +0200 (Daylight)".
However, the entry in a calendar multiget from the inbox had the original buggy TZID.
After the fix, able to propose a new time using ZWC and the CounterAppointmentRequest uses "Europe/Berlin" rather than the buggy timezone
https://reviewboard.eng.zimbra.com/r/3313/

https://bugzilla.zimbra.com/show_bug.cgi?id=86249

Copied from Perforce
 Change: 531396