(split)Bug: 51541
TP user story: https://telligent.tpondemand.com/entity/88582

Added "is:tomeonly" search query operator. The operator works by constraining the "recipients" column of the mail_item table to the requesting mailbox's account name and aliases. The name to be filtered on is fed through the same ParsedAddress.getSortString() method that is invoked when inserting recipients data. Although the "is:tome" query searches the "to" field of the index, "tomeonly"  has to use the database because it is impractical to search the index for documents containing ONLY the user's account name in to "to" field, given how we tokenize addresses. 

It should be noted that I did not update the DbSearchConstraints.Leaf.toQueryString() method to encode this new operator into a query string for remote execution, since "is:tomeonly", by definition, only applies to the user's mailbox.

ToMeOnlyQuery.java:
 - Query subclass that compiles to a DbQueryOperation with the specified recipients constraint

BuiltInQuery.java:
 - mapping "tomeonly" to an instance of ToMeOnlyQuery

DbQueryOperation.java:
 - added addRecipients() method

DbSearchConstraints.java:
 - added addRecipients() method, and corresponding "recipients" and "excludeRecipients" fields

DbSearch.java:
 - updated encodeConstraints() method to encode the "recipients" and "excludeRecipients" fields

TestIsToMeOnly.java:
 - unit test verifying functionality of "is:tomeonly", "-is:tomeonly", and several more compound queries involving the operators

https://bugzilla.zimbra.com/show_bug.cgi?id=51541

Copied from Perforce
 Change: 531397