(split)bug: 95933
Reinstate old checkpoint increments; except when always on 'mode' is enabled. This enables the system to keep operating with data which was created in 8.0.x and which is therefore expected to have gaps between mail_item_checkpoint and max(mail_item.id); as well as reducing the number of DB writes for normal (non 'alwayson') operation.

https://bugzilla.zimbra.com/show_bug.cgi?id=95933

Copied from Perforce
 Change: 531664