(split)bug: 96870
Check isAlwaysOn() each time we lookup the checkpoint increments; so that we can adjust if a server joins a cluster while running. Eventually we may want to create a 'join cluster' event so call sites can listen for changes rather than checking on every access.

Increment the checkpoints when a mailbox is restored on a server running in AlwaysOn mode. We cannot easily determine the checkpoint increments used on the source system at the time the backup was created, so we err in favor of the more common case moving a mailbox between a standalone node and always on cluster. It is harmless if we happen to increment the checkpoint where it is not actually necessary.

https://bugzilla.zimbra.com/show_bug.cgi?id=96870

Copied from Perforce
 Change: 531947