(split)Bug: 95991
Keeping track of commits using SOLR's postCommit event is a good approach, but it currently runs into issues when more than one document is added to the index in a single unit test. To properly handle this, we need to keep track of how many update requests have been sent to the embedded solr index, and block until that many postCommit events have fired. Additionally, Mailbox.indexDeferredItems() should be called in waitUntilIndexingCompleted(), since in some cases messages will not be committed in time for them to be visible in the index when necessary.

ZimbraCommitListener.java:
 - Changed the method signature of waituntilCommitFinished() to include the number of permits to request. This should correspond to the number of times the postCommit event is expected to be called.

EmbeddedSolrIndex.java:
 - Added an inner class UpdateCounter that is used to keep track of the number of updates sent to embedded solr server.
 - Added a call to update this counter in the add() method.

MailboxTestUtil.java:
 - Added a call to mbox.indexDeferredItems() in waitUntilIndexingCompleted().
 - When calling CommitLock.waitUntilCommitFinished(), passing the number of updates from UpdateCounter. This also resets the counter.
 - Catching IOException in cleanupAllIndexStores(), which can be thrown when an index directory cannot be deleted.

MailboxTest.java:
 - Removed unnecessary calls to waitUntilIndexingCompleted().

https://bugzilla.zimbra.com/show_bug.cgi?id=95991

Copied from Perforce
 Change: 532054