(split)Bug 55152 - Prevent access to Zimbra if source IP is not within sanctioned IP address range behind VPN
Correcting the earlier implementation which was incorrect. 
With these changes, Nginx fowards the client IP to Nginxlookup servlet, the servlet does the IP matching (i.e checks whether the client IP falls under one of the IP ranges configured under zimbraDomainAllowedIPs). If it does not match, sends an error message back to the proxy & the client gets a 403 Acess Forbidden. If it does match, client is able to login as usual.
1. Renaming zimbraReverseProxyDomainAllowedIPs to zimbraDomainAllowedIPs as this is not a proxy attibute anymore
2. IP matching code within NginxLookupHandler
3. Changes in main/pom.xml & ZimbraNginxLookup/pom.xml to include the apache commons-net dependency
4. Changes in the nginx code to parse this specific access denied error message, bypass selecting any upstream by IPHASH & throw a 403 to the user
5. Removing the unneeded allow/deny directives from the vhost nginx templates

https://bugzilla.zimbra.com/show_bug.cgi?id=55152

Copied from Perforce
 Change: 533745