(split)bug: 97612
TestJaxb.testBrowseRequestAttachmentsBadRegex - make sure there is at least 1 attachment.
TestJaxb.testAcceptSeriesDeclineInstance
    Sometimes when procesing hits returned from the index, the associated cached mail item still has 0 as the indexId.
    Fix is to use the id if the indexid is one of the special values used when it isn't a valid indexid.
    Note that according to MailItem.getIndexId()
Updated comment for MailItem.getIndexId() with more details on possible values (that was where I got the idea it was possible that indexId could be valid but NOT the same as the id.
Testing:  Ran TestJaxb multiple times with debug statement logging when indexId has been over-ridden with id.
          Tests always pass.  The logging is present about about half the time.

https://reviewboard.eng.zimbra.com/r/3961/

https://bugzilla.zimbra.com/show_bug.cgi?id=97612

Copied from Perforce
 Change: 536356