(split)Bug 80944 - zimbraIPMode "both" does not work correctly for proxy when zimbraVirtualHostname is used 1. In case of zimbraIPMode=both, need just one listen directive with the ipv6 addr & ipv6only option as 'off' (This change in behavior is because of change in default value for 'ipv6only' from off to on in nginx 1.3.4) Ref: http://codinginthetrenches.com/2013/12/14/nginx-ipv6only-setting-gotcha/
ipv6only=on|off
this parameter (0.7.42) determines (via the IPV6_V6ONLY socket option) whether an IPv6 socket listening on a wildcard address [::] will accept only IPv6 connections or both IPv6 and IPv4 connections. This parameter is turned on by default. It can only be set once on start.
Prior to version 1.3.4, if this parameter was omitted then the operating system’s settings were in effect for the socket.

2. No need of explicit 'ipv6only=on' in case of zimbraIPMode=ipv6

https://bugzilla.zimbra.com/show_bug.cgi?id=80944

Copied from Perforce
 Change: 536560