(split)Bug: 95987
Apple Delegate creation via CalDAV.
Added tests.
Fixed a couple of issues noticed whilst doing that testing.
AbstractCalendarProxy test for whether was read-only didn't work if URL was "/" terminated.  Use different method.
Prefer "/" termination for those URLs because they are collections.

Seen a few more href's without %40 encoding instead of @
When looking for a resource with one of the proxy URLs when zimbraPrefAppleIcalDelegationEnabled=FALSE, we now do a 404 NOT FOUND instead of returning a User resource.
This results in a slightly better error in Mac Calendar when you hit the "Edit" button instead of an unexpected 424.  You get "The calendar was not found on the server".

https://reviewboard.eng.zimbra.com/r/4050/

https://bugzilla.zimbra.com/show_bug.cgi?id=95987

Copied from Perforce
 Change: 536663