(split)Bug 97944 - Support Redis Cluster
Construct a JedisCluster singleton when possible (when the Redis URLs point to a viable cluster). Provide and use a Redis Cluster based mailbox pub-sub adapter, and shared delivery coordinator, to be used instead of the default Redis adapter, when a cluster is available. This fixes session notifications, and delivery to multiple recipients, when Redis Cluster is in use.

https://bugzilla.zimbra.com/show_bug.cgi?id=97944

Copied from Perforce
 Change: 537619