(split)Bug 85257 - Distributed Mailbox Lock management
Provide scaffolding for a Redis-based MailboxLock adapter, which is the Redis-based replacement of the ZooKeeper distributed lock. The Redis-based MailboxLock adapter is not enabled yet, and is an incomplete implementation. But it is being committed for peer review. A final implementation may switch to using RedissonLock, but that class also has test coverage that does not pass yet.

As of this change, mailbox locking is still hard-coded against the LocalMailboxLock adapter, which is the legacy in-process Java-based concurrency management.

https://bugzilla.zimbra.com/show_bug.cgi?id=85257

Copied from Perforce
 Change: 540076