Bug 98176 - FirstName and LastName of fullName is joined as fixed order in addMatchedContacts method
[Summary]
Change the server to return additional attributes in the autocomplete response so that the client can calculate FileAs the same way it does for contacts.

[Problem Description]
While message was composed, Web client (Ajax and standard HTML)
displayed a list of addresses incrementally for the address auto-
completion.  The candidate addresses were taken from the user's 
contact list, address group, distribution list, or GAL, and each 
candidate address was displayed with a label string which was
supposed to be the same label string displayed at the list in the 
Contact app.

Bug it was found that some auto-completion candidate addresses were 
displayed with different label string than the one in the Contact
list: Even if the label string was displayed "lastName firstName" 
in the Contact list, it was shown as "firstName lastName" in the 
auto-complete list.

Since the label string on the auto-completion address may be copied
to the To/Cc/Bcc fields as a display name and embedded on the email
header, the label string should have obeyed the display order rule 
which the end user specified with the "FileAs" attribute in the 
Contact app.

[Root Cause]
The auto-completion candidate addresses were searched on the server 
side through the AutoCompleteRequest SOAP request.  The mailstore 
returned a list of candidate addresses with a label string which was 
generated on the mailstore side based on the following rule:
  - "fullName" attribute if the "fullName" was not null
  - Combination of "firstName", followed by "LastName"
    if the "fullName" is null

After ZCS 7 (as long as I tested), the ZCS client sets "fullName".
But previous Zimbra client, such as 5.0.16 HTML client did not set
the "fullName" attribute at all, and 5.0.16 Ajax client always set
"firstName lastName" for "fullName" (it ignored the "FileAs").
In the end, when the system was upgraded, some Contact item still
had incorrect "fullName" attribute on the new system (those corrupt 
contact item will never be corrected unless the end user modifies 
it explicitly).
As a result, the AutoComplete returned the incorrect label string.

On the other hand, the label string which is displayed in the 
Contact list is generated by the ZmContact.js#computeFileAs; this 
function assembles a label string using "firstName", "lastName", 
"nickName", "companyName", and "fullName", in accordance with the 
"FileAs" attribute, and corresponding label pattern specified in 
the localized ZmMsg properties file.

The label string for the Contact list was composed on the client
side, whereas the label for the auto-complete candidate addresse
was made on the server side with different generating rule.

[Fix Description]
Basically, this fix bunches up together the logic of generating the 
label string of the auto-completion address to the client side 
because the format of the label string is defined on the client side.

(1) Ajax Client
If the auto-complete candidate addresses are member of the Contact 
list, ZmAutocomplete.js#ZmAutocompleteMatch() will lookup the id of 
the candidate contact item to the Contact list, and get the Contact 
item object to pass the ZmContact.js#computeFileAs() so that the same 
label string as in the Contact list is generated.

(2) Mailstore (AutoCompleteRequest/Response)
The AutoCompleteRequest/Response is modified to send back "firstName",
"middleName", "lastName", "fullname", "nickname", "companyName", and
"fileAs" along with other current elements.  

(3) HTML Client
Before HTML client displays the candidate addresses on the screen, 
JavaScript formatNameLabel() function defined in the autoComplete.tag 
will be called to re-format the label string from the "firstName", 
"lastName", "fullName" etc. which are a part of the payload of the 
AutoCompleteResponse.

[Affected area]
Ajax and HTML Client: message composing window/auto complete

[Configuration changes]
None

[LDAP Schema changes]
None

[API changes]
SOAP request: Parameter of the AutoCompleteResponse.
   Added parameters:
      first - (optional) first name of the matched contact item.
      middle - (optional) middle name of the matched contact item.
      last - (optional) last name of the matched contact item.
      full - (optional) full name of the matched contact item.
             In most every case, this string is the same as the 
             display name which is the head part of the "email" 
             parameter of the AutoCompleteResponse.
      company - (optional) company name of the matched contact item.
      fileas - (optional) FileAs parameter of the matched contact item.

[Log changes]
None

[Manual Test]
(Test A) Created a contact item with/without fullName, and FileAs area
  in the "metadata" field of the mail_item table for the target user, 
  and verified that the label string of the auto-complete candidate 
  addresses was displayed same as that in the Contact list.
(Test B) Verified the test A on the HTML client and Ajax client.
(Test C) Performed test A with various FileAs value, such as last-first,
  first-last, etc. and verified that label string of the auto-complete 
  address was displayed same as that in the Contact list.
(Test D) Performed the test A on the Japanese locale, and English locale,
  and verified that the label string of the auto-complete address was 
  complied with the locale, and displayed same as that in the Contact 
  list.
(Test E) Created a group, distribution list, and verified that auto-
  complete functionality worked.
(Test F) Created another account, performed the auto-complete request,
  and verified that the auto-complete functionality worked.

https://bugzilla.zimbra.com/show_bug.cgi?id=98176

Copied from Perforce
 Change: 540219