(split)Bug 98724 - When always on mode is enabled, folder and tag caches are rebuilt for every transaction
Preliminary fix to remove folders and tags cache reloading, providing some immediate relief. This stateless data access strategy from Always On phase 1 was premature, and doesn't work right because the data structures that are cached and the structures that are loaded from the DB are not the same, or even 1-to-1 in call counts.

To finish fixing this bug for this release, we'll use pub/sub notifications to signal that the caches should be reloaded. That will depend on bug 98684 getting fixed first, getting some pub/sub plumbing in place first.

https://bugzilla.zimbra.com/show_bug.cgi?id=98724

Copied from Perforce
 Change: 541779