(split)Bug 98423 Improve error logging done nginxlookup client
This is to enhance error logging since we think the Admin client Consul lookups are currently wrong and sending nginx-lookup requests to the wrong kind of endpoints. Currently the code throws a NPE.

Also improve some generics in the code to reduce warnings. And abstracted the memcached-based route cache into an interface+adapter for general clean-up.

https://bugzilla.zimbra.com/show_bug.cgi?id=98423

Copied from Perforce
 Change: 543159