(split)Bug 98423 - Web client is caching routes to mailstore and nginx-lookup
Don't forward soap requests to a remote server if we're on a running server where the request can be handled internally. This fixes webapp requests being soap-forwarded to mailstore-a, and mailstore-a forwarding it to mailstore-any instead of handling it internally.

This changelist results in the first successful completion of a mailstore failover test. The proxy and web client front-ends switch from a Mailstore-A to a Mailstore-B after a controlled shutdown of Mailstore-A, followed by a reload of the web client page.

https://bugzilla.zimbra.com/show_bug.cgi?id=98423

Copied from Perforce
 Change: 543750