(split)Bug 98742 - When Always On is enabled, folder and tag caches are rebuilt every transaction
This bug seems to be good to go as-is, after a final round of testing. Just committing some more cache-related clean-up here. The memcached FoldersAndTagsCache adapter had some duplicate code. And a FoldersAndTagsCacheMailboxListener is provided, but won't be needed until a future release where the mailstore becomes stateless.

https://bugzilla.zimbra.com/show_bug.cgi?id=98742

Copied from Perforce
 Change: 543988