(split)Bug 99343 - Halfwidth Katakana word is not indexed correctly in lucene: "voiced" and "semi-voiced" sound marks are treated incrrectly
[Problem Description]
If an indexable word, such as message text, contact item, calendar appointment, etc. includes a "voiced" and/or "semi-voiced" sound mark of the Japanese half-width Katakana letter, that word is not indexed correctly; as a result, such word will not be searched correctly.

For example, if the last name of a contact item, or message body contains ｺﾞｼﾞﾗ, it is mistakenly treated as コ゛シ゛ラ, it was supposed to be tokenized as ゴジラ.  

[Root Cause]
When the text (message body, contact item, etc.) was tokenized, the NormalizeCharFilter.normalize() method was called to normalize one character by one character.  Regarding the Japanese letters, the method converted the half-width Katakana character to the full-width, as well as converted the full-width Katakana to the full-width Hiragana (in the end, the Japanese Katakana letter is tokenized as Hiragana).
Typical normalized flow is:
  Half-width Katakana --> Full-width Katakana --> Full-width Hiragana   
By the way, some Japanese half-width letters followed by a "voiced" or "semi-voiced" sound marks should be mapped two characters to one full-width character.  For example, the ｺ followed by the voiced sound mark of ﾞ (Note 1) should have been mapped to "ゴ".  But because the normalize() method took only one letter at a time, it returned the こ first (first converted to the full-width Katakana コ then full-width Hiragana こ), then returned the full-width of the voiced sound mark ゛.  On the other hand, the full-width character with voiced sound mark ゴ was treated as the corresponding Hiragana letter ご.  As a result, the half-width Katakana word was indexed with こ゛(2 characters), whereas the search keyword may be issued with ご (1 character), and the item with half-width voiced sound marks would be never searched (Note 2).

(Note 1) the voice sound mark looks like a double-quotation mark, but it is always attached with a particular kind of consonant character.  The semi-voiced sound mark looks a small circle (like a tiny circle for the degree of temperature), but same rule of the voice sound mark should be applied too.
(Note 2) Of course, if the search keyword was consisted of the half-width letter, then the item would be searched.  But usually people don't search with half-width letters.

[Fix Description]
Introduced a text based mapping table file to convert the half-width Katakana characters followed by the voiced/semi-voiced sound mark by solr.MappingCharFilterFactory before the NormalizeCharFilter.normalized() method is called.

I tried to fix the normalize() method, but because it is called not only from the Zimbra's class but the native Solr/Lucene code, I chose to go with solr.MappingCharFilterFactory.

[Affected Area]
SearchRequest: 
* User can find any half-width Katakana words in the text items (message body, etc.) and contact items which is newly indexed with any half/full-width Katakana/Hiragana search keyword.
* User can find any full-width Katakana/Hiragana words with any half-width Katakana search keyword.
* The existing index data is not affected by this change; which means that the existing index with half-width characters with voiced/semi-voiced sound marks will not be able to be searched until the data is re-indexed.

[Configuration Changes]
None

[LDAP Schema Changes]
None

[API Changes]
None

[Log Changes]
None

https://bugzilla.zimbra.com/show_bug.cgi?id=99343

Copied from Perforce
 Change: 544299