(split)Bug 84987 - Messages cannot be removed when zimbraMailEmptyFolderBatchSize is 0
[Problem Description]
When the global/server config 'zimbraMailEmptyFolderBatchThreshold' is explicitly set to 0, "Empty Folder" operation does not work and all the messages under the specified folder remained.

The end user can invoke the "Empty Folder" operation, and they even see that the message list is temporarily deleted.  But if you look into the mailbox (by zmmailbox or directly SELECT the DB), you will see the messages are still in the specific folder, and once the folder is reloaded, the message list is reappeared.

[Root Cause]
When a list of deleting candidate messages is retrieved from DB, the 'LIMIT' clause will be added under a certain condition; the LIMIT clause should have been added when the zimbraMailEmptyFolderBatchSize is set by the number which is more than the value of 0, but LIMIT clause was added always when the zimbraMailEmptyFolderBatchSize is set some value, even 0. As a result, the DB returned an empty message list.

[Fix Description]
Skip adding an LIMIT clause when the zimbraMailEmptyFolderBatchSize is set to 0.

[Affected Area]
Ajax client: Mail app - 'Empty Folder'

[Configuration Changes]
None

[LDAP Schema Changes]
None

[API Changes]
None

[Log Changes]
None

[Manual Test]
Verified the steps 1-5 described in this bug.

https://bugzilla.zimbra.com/show_bug.cgi?id=84987

Copied from Perforce
 Change: 544724