(split)Bug:98916
VCard with large attachment causes server problems.
Changed VCard parsing to use StringBuilder instead of appending each line to a string!!!
With this change, DAV POST/PUT requests are responded to within a few seconds instead of minutes (due to not causing the garbage collector to have to work overtime) and therefore BusyContacts doesn't get a timeout, which caused it to retry.

Tested with attachments of around 3MB, 6.5MB, just more than 7MB and 18.9MB.  With a patched proper installation of 8.7.0.BETA1.1204.UBUNTU12.64, the first 2 succeed.  The other 2 do not and BusyContacts either reports that it failed and removes the card or reports that it failed and that the attachments have been removed. (actually, latter are because BusyContacts spots they are too big and drops them before sending a POST/PUT).

Using curl with a VCARD containing abase64 encoding of an 8.6MB file, I get this error:
HTTP/1.1 413 Request Entity Too Large
Corresponding mailbox.log info:
     FileUploadServlet - Exceeded maximum upload size of 10485760 bytes
     dav - sending http error 413 because: can't save upload

With a dev build, they all report failure on my mac due to a problem with the mysql driver failing the insert.
https://reviewboard.eng.zimbra.com/r/4968/

https://bugzilla.zimbra.com/show_bug.cgi?id=98916

Copied from Perforce
 Change: 544887