(split)Bug 78390 - The filter condition "Address in To or Cc" does not work
[Problem Description]
If the filter is set a rule with "Address in" - "To or Cc" - "in" - "Contacts", the message whose To or Cc is in the contact list should be filtered, but the filter does not work at all (none of them are filtered).

[Root Cause]
For example, ZCS generated the SIEVE rule text for such filter:
-----
require ["fileinto", "reject", "tag", "flag"];

# sample1 
if anyof (addressbook :in "TO,CC") {
    flag "flagged";
    stop;
}
-----
This rule filtered if the email address specified by the field name "TO,CC" in the message header is in the contact list; the syntax of the condition part was wrong.  

[Fix Description]
Fixed SoapToSieve.handleTest() for handling the ModifyFilterRulesRequest that the header field name is listed within the [] "(addressbook :in ["TO","CC"])", and SieveVisitor.acceptTest() for the GetFilterRulesRequest so that all the header fields are gathered.

[Affected Area]
Get or set filter rule of "Address in" - "To or Cc" - "in" - "Contacts" via web client or command line.

[Configuration Changes]
None

[LDAP Schema Changes]
None

[API Changes]
None

[Log Changes]
None

[Manual Test]
1/ Login to the Ajax web client and open the Filters page in the Preferences tab.
2/ Click "Create Filter", and set the following values:
   Filter Name: test
   If "Any" of the following conditions are met:
   "Address in" - "To or Cc" - "in" - "Contacts"
   "Mark as" - "Flagged"
   And save it.
3/ Go to the Contacts tab and create a contact item for your own email address.
4/ Compose a message and set To to yourself, and send it.
5/ Update the message list and verify that the message sent in the Step 4 has a flag.

6/ Compose another message and set Cc to yourself.
7/ Update the message list and verify that the message sent in the Step 6 has a flag too.

8/ Go to the Filters page in the Preferences tab again.
9/ Open the filter "test" (created in the step 2) and verify that the condition set has not been modified.

10/ Check the sieve filter text via zmprov
  $ zmprov ga user1 zimbraMailSieveScript
  # name user1@ykomiyam.local
  zimbraMailSieveScript: require ["fileinto", "reject", "tag", "flag"];

  # test 
  if anyof (addressbook :in ["TO","CC"]) {
      flag "flagged";
      stop;
  }

https://bugzilla.zimbra.com/show_bug.cgi?id=78390

Copied from Perforce
 Change: 545075